Feat/custom nightbot replacement (#2)
* Untested Attempt

The bot currently should send a message to the chat, then send a whisper to the user that fills them in on some further details. I can't test it personally because twitch doesn't let you whisper yourself but I'll update with test changes.

* Untested clean command

I often find myself in need of a way to clear a user's messages after breaking the rules, and forget the twitch integrated way of doing so, hence this command existing.

* Gutenberg Command Post-Test

A command that explains to the user where I find my public domain books. Intended to send messages, but the API was having trouble with that.

* Discord Announcement

Added a command that announces a discord based on the link provided in the properties file. If there's no link provided in the file, it has a message saying there's no discord.

* Friend Code Outputter

Based on some items from the properties file, this outputs a given friendcode. In theory I could make this work for any item based on the arguments passed to the command, but for now we'll leave it limited like this.

* FC Modularity

Since the friendcode command essentially grabs some value from a file, I rolled a promo command into it's functionality in the process of making it entirely modular. Does it not make much sense to say Promo switch to get the switch friendcode? yeah, but nobody's going to use it like that, they'll use the command that makes sense.

* Completed Custom Commands

My custom nightbot commands now have been implemented, so this branch should get merged into dev, and I can move on to implementing the default parts of nightbot I use.